Bug_Fix for the forking issue
This should fix the issue of a forking BAZO-Blockchain.
Implemented a stash for all received blocks. After a rollback now the
miner first checks those blocks before requesting one from the network.
the Block request to the network is now done with a broadcast request,
such that the possibility of a valid answer is maximised.
When a fork occurs, the state of the fork-brnaches can quickly diverge.
This results in transactions listed as invalid, despite they are valid
after a rollback to the correct chain. Now we do not discard them
automatically but write them in a special stash, from where they can be
fetched after the rollback. This results in less network requests and a
higher possibility to get a valid answer. Also here the Tx request is
now a broadcast. This also ensures for the case when a transaction is
not received before. With only one random selected miner the miner may
has bad luck and does not get the transaction. Then the block is invalid
and this should be prevented.